+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * added `write_json_to_file`
 * added `read_json_from_file`
 * added `write_json_to_file_if_different`
+* added `satella.distutils`